requirements.txt
slack-bolt
pandas
matplotlib
openpyxl
python-dotenv
schedule
